Add tests for line-ending normalisation
This gets some tests set up. For now, only the line-ending normalisation is tested.
