+++ tests/pch.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "../foo_uie_console/main.h"
