+++ vcpkg.json
@@ -2,8 +2,12 @@
   "name": "console-panel",
   "version-string": "git",
   "builtin-baseline": "6062c8fe02d38b10d5291411bc235ae4d02cfe6c",
-  "dependencies": ["fmt", "ms-gsl", "range-v3", "wil"],
+  "dependencies": ["fmt", "ms-gsl", "range-v3", "wil", "doctest"],
   "overrides": [
+    {
+      "name": "doctest",
+      "version": "2.4.12"
+    },
     {
       "name": "fmt",
       "version": "10.2.1#2"
